move save_proofing_components to new DocumentCaptureConcern
